Implement custom OpenAPI schema generation and remove unused code; add version parameter to all endpoints
