+++ packages/api/src/_lib/shared.py
@@ -2,8 +2,7 @@ import logging
 from datetime import datetime
 from enum import Enum
 
-from fastapi import FastAPI, Header, HTTPException, Request
-from fastapi.openapi.utils import get_openapi
+from fastapi import Header, HTTPException, Request
 
 
 logger = logging.getLogger(__name__)
@@ -52,63 +51,3 @@ def add_version_headers(response, current_version: ApiVersion):
 
     if current_version.preview:
         response.headers["X-Version-Status"] = "Preview"
-
-
-def custom_openapi(app: FastAPI):
-    def custom_openapi_func():
-        try:
-            logger.debug("Generating custom OpenAPI schema")
-            if app.openapi_schema:
-                logger.debug("Returning cached OpenAPI schema")
-                return app.openapi_schema
-
-            openapi_schema = get_openapi(
-                title="Your API",
-                version="1.0.0",
-                description="Your API description",
-                routes=app.routes,
-            )
-
-            logger.debug("Adding version parameter to all endpoints")
-            # Define the version parameter once at the components level
-            openapi_schema["components"] = openapi_schema.get("components", {})
-            openapi_schema["components"]["parameters"] = {
-                "ApiVersionHeader": {
-                    "name": "x-api-version",
-                    "in": "header",
-                    "required": True,
-                    "schema": {
-                        "type": "string",
-                        "enum": [v.value for v in ApiVersion],
-                        "default": ApiVersion.LATEST.value,
-                    },
-                    "description": "API Version to use",
-                }
-            }
-
-            # Reference the parameter in all paths
-            for path in openapi_schema["paths"].values():
-                for method in path.values():
-                    # Remove any existing x-api-version parameters
-                    if "parameters" in method:
-                        method["parameters"] = [
-                            param
-                            for param in method["parameters"]
-                            if param.get("name", "").lower() != "x-api-version"
-                        ]
-
-                    # Add the reference to our component parameter
-                    method["parameters"] = [
-                        {"$ref": "#/components/parameters/ApiVersionHeader"}
-                    ] + (method.get("parameters", []))
-
-            # Modify the servers to include the /api prefix
-            openapi_schema["servers"] = [{"url": "/api"}]
-            app.openapi_schema = openapi_schema
-
-            return app.openapi_schema
-        except Exception as e:
-            logger.error(f"Error generating OpenAPI schema: {str(e)}")
-            raise
-
-    return custom_openapi_func
